requirements.txt
appdirs==1.4.4
args==0.1.0
clint==0.5.1
jedi==0.18.1
numpy==1.22.3
parso==0.8.3
Pillow==9.1.0
prompt-toolkit==3.0.29
ptpython==3.0.20
Pygments==2.12.0
wcwidth==0.2.5
